update related clasee to use Formatter
