+++ src/seedu/addressbook/ui/Formatter.java
@@ -1,5 +1,7 @@
 package seedu.addressbook.ui;
 
+import java.util.List;
+
 public class Formatter {
 
 
@@ -31,20 +33,39 @@ public class Formatter {
      */
     private static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
 
-    public String getDivider() {
+    public static String getDivider() {
         return DIVIDER;
     }
 
-    public String getLinePrefix() {
+    public static String getLinePrefix() {
         return LINE_PREFIX;
     }
 
-    public String getLS() {
+    public static String getLS() {
         return LS;
     }
 
-    public String getCommentLineFormatRegex() {
+    public static String getCommentLineFormatRegex() {
         return COMMENT_LINE_FORMAT_REGEX;
     }
 
+    /** Formats a list of strings as a viewable indexed list. */
+    public static String getIndexedListForViewing(List<String> listItems) {
+        final StringBuilder formatted = new StringBuilder();
+        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
+        for (String listItem : listItems) {
+            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
+            displayIndex++;
+        }
+        return formatted.toString();
+    }
+
+    /**
+     * Formats a string as a viewable indexed list item.
+     *
+     * @param visibleIndex visible index for this listing
+     */
+    private static String getIndexedListItem(int visibleIndex, String listItem) {
+        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
+    }
 }
